+++ test/test_listen_callback.cpp
@@ -213,7 +213,7 @@ int SrtTestListenCallback(void* opaq, SRTSOCKET ns, int hsversion, const struct 
 
 #if SRT_ENABLE_ENCRYPTION
     cerr << "TEST: Setting password '" << exp_pw << "' as per user '" << username << "'\n";
-    EXPECT_EQ(srt_setsockflag(ns, SRTO_PASSPHRASE, exp_pw.c_str(), exp_pw.size()), SRT_SUCCESS);
+    EXPECT_EQ(srt_setsockflag(ns, SRTO_PASSPHRASE, exp_pw.c_str(), (int)exp_pw.size()), SRT_SUCCESS);
 #endif
 
     // Checking that SRTO_RCVLATENCY (PRE option) can be altered in the listener callback.
@@ -241,9 +241,9 @@ TEST_F(ListenerCallback, SecureSuccess)
     string username_spec = "#!::u=admin";
     string password = "thelocalmanager";
 
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), username_spec.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), (int)username_spec.size()), -1);
 #if SRT_ENABLE_ENCRYPTION
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), password.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), (int)password.size()), -1);
 #endif
 
     // EXPECTED RESULT: connected successfully
@@ -258,8 +258,8 @@ TEST_F(ListenerCallback, FauxPass)
     string username_spec = "#!::u=admin";
     string password = "thelokalmanager"; // (typo :D)
 
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), username_spec.size()), -1);
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), password.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), (int)username_spec.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), (int)password.size()), -1);
 
     // EXPECTED RESULT: connection rejected
     EXPECT_EQ(srt_connect(client_sock, psa, sizeof sa), SRT_ERROR);
@@ -273,9 +273,9 @@ TEST_F(ListenerCallback, FauxUser)
     string username_spec = "#!::u=haivision";
     string password = "thelocalmanager"; // (typo :D)
 
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), username_spec.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), (int)username_spec.size()), -1);
 #if SRT_ENABLE_ENCRYPTION
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), password.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), (int)password.size()), -1);
 #endif
 
     // EXPECTED RESULT: connection rejected
@@ -289,9 +289,9 @@ TEST_F(ListenerCallback, FauxSyntax)
     string username_spec = "#!::r=mystream,t=publish"; // No 'u' key specified
     string password = "thelocalmanager";
 
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), username_spec.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_STREAMID, username_spec.c_str(), (int)username_spec.size()), -1);
 #if SRT_ENABLE_ENCRYPTION
-    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), password.size()), -1);
+    ASSERT_NE(srt_setsockflag(client_sock, SRTO_PASSPHRASE, password.c_str(), (int)password.size()), -1);
 #endif
 
     // EXPECTED RESULT: connection rejected
